DF-11859 set "flag warning as error" for android; fix errors
Former-commit-id: 43978a44ca7180cac5895a257b34121c9cf3cfc9

+++ Sources/Internal/UI/Private/Android/MovieViewControlAndroid.h
@@ -77,7 +77,7 @@ namespace DAVA
 class JniMovieViewControl
 {
 public:
-    JniMovieViewControl(uint32 id);
+    JniMovieViewControl(uintptr_t id);
     void Initialize(const Rect& rect);
     void Uninitialize();
